fixed playerinput handler, always read my descrip
so the way the old input handler was doing it, only one input handler
could be active at a time, so we couldn't create two instances one for
each player.

I changed it so it figures out which player it is handling then the
update function checks the keys. So player 1 and 2 create their own
handler, and they have different keys. read other post to see new
controls