Cleaned up hazard test screen, but...
Kevin and I tried to get the HazardScreen to extend Screen instead of
implementing it, but we ran into a couple of snags. On the bright side,
we fixed an issue with the AssetManager which would have caused it to
not-load if LoadingScreen was bypassed somehow, and we did clean up a
lot of redundant code. Unfortunately, we ran into a camera bug with
HazardScreen (other screens work fine), so we're going to wait until we
can talk to Ed R.