added a title background that's 1024.
I recall Kevin mentioning something about powers of 2, so I added a
title background that's a power of 2. I thought about using it, so that
it sort of moves the title to the middle for the presentation, but I
didn't want to implement it without letting the team know first that
it's a hacky fix.