More work on sounds. As usual, I found a lot more work to be done than I thought there was.
I added a function called postLoad() to Entity that should be
called during build(). It starts the entity's "idle" sound. Need
to check that it's being called in all the necessary builders.

Also added back the jump sound to player (will replace with
Victor's sound soon) as well as a dummy sound.