Fixed Stewart's decal origin problem and the multiple headjump sound problem.
Blocked on idle sounds because I currently have no way to
get the screen space coordinates of a sprite. I've tried
SpriteBatch.getProjectionMatrix but that gave me nonsense coordinates,
so I'm looking for another option.

Also would like to get started on collision sounds, but since that will
require messing with the ContactManager, I'm going to wait for Ed B and
Jen to finish adding metrics to the collision code so we don't have to
re-do it.